Arm backend: Move aot_arm_compiler to backends/arm/scripts/
Move aot_arm_compiler.py to from examples/arm/ to backends/arm/scripts/
for the following reasons:
 - It behaves more like a script/tool than an example. An example is
   often focused and concise which the aot_arm_compiler is not.
 - It will now be installed with the executorch pip package, meaning it
   can be run like this `python -m
   executorch.backends.arm.scripts.aot_arm_compiler` instead of needing
   to be executed via its file path from the executorch source tree.
 - Since it is now installed with the pip package, functions in it can
   be imported elsewhere, which is a useful utility we will utilize.

With the migration, a number of MyPy and Flake8 errors arose. This
patch fixes those errors, meaning that the new version of
aot_arm_compiler is not a direct copy of its former version.

For backwards capability, the old file is still kept in examples/arm/;
it simply invokes the main function of the new version.
